moving matrix embedding to stand-alone file
